app/src/main/java/com/vinodpatildev/dailynews/data/api/NewsApiService.kt
package com.vinodpatildev.dailynews.data.api

import com.vinodpatildev.dailynews.BuildConfig
import com.vinodpatildev.dailynews.data.model.ApiResponse
import retrofit2.Response
import retrofit2.http.GET
import retrofit2.http.Query

interface NewsApiService {
    @GET("v2/top-headlines")
    suspend fun getTopNewsHeadlines(
        @Query("country")
        country:String,
        @Query("page")
        page:Int,
        @Query("apiKey")
        apiKey:String = BuildConfig.NEWS_API_KEY
    ): Response<ApiResponse>

    @GET("v2/top-headlines")
    suspend fun getSearchedNewsHeadlines(
        @Query("country")
        country:String,
        @Query("q")
        searchQuery:String,
        @Query("page")
        page:Int,
        @Query("apiKey")
        apiKey:String = BuildConfig.NEWS_API_KEY
    ): Response<ApiResponse>


}